gpt4all-training/requirements.txt
accelerate
datasets
einops
torchmetrics
evaluate
transformers>=4.28.0
wandb
peft
nodelist-inflator
deepspeed
sentencepiece
jsonlines
nomic
scikit-learn
matplotlib
absl-py
rouge-score
nltk